labels inside of for and while loops were not recognized
recursion through node array needed to be added as nodes in ForNodes and WhileNodes are nested. function 'define_labels' finds all labels first after parsing. Nesting of blocks in loop nodes causes labels inside the blocks to be looked over during the select statement. Recersive function 'label_recur' was added to handle this. A more elegant solution is surely possible, as WhileNode and ForNode are explicitely cited. Getter function for @block class variable are added. attr_reader could not be used due to conflict with block def's in each class.